Fix critical and major sonar violations
Issue-ID: AAI-342

+++ sparkybe-onap-service/src/main/java/org/onap/aai/sparky/analytics/HistogramSampler.java
@@ -173,7 +173,7 @@ public final class HistogramSampler {
     this.minValue = -1;
     this.maxValue = 0;
     initializeHistogramBins(numBins);
-    this.stepSize = (maxValue / numBins);
+    this.stepSize = ((double) maxValue / (double) numBins);
     clear();
   }
